# web/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Despesas Pessoais",
        "description": "API RESTful para gerenciamento de despesas pessoais (Yii2 + JWT).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Ambiente local (Docker)"
        }
    ],
    "paths": {
...
        "/expenses": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Lista as despesas do usuário (com filtros, ordenação e paginação)",
                "description": "GET /expenses\nLista as despesas do usuário com filtros (category, month, year),\nordenação (sort=asc|desc) e paginação (page, per_page).",
                "operationId": "0ed5f1ce2af9c96f0c2f8b1fa1c7d70d",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "alimentacao",
                                "transporte",
                                "lazer"
                            ]
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2026
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Exige o parâmetro year",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 12,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de despesas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Expense"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Parâmetros de busca inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Cria uma nova despesa",
                "description": "POST /expenses\nCria uma nova despesa.",
                "operationId": "a9a2b52de18c3587f59df5d66680a812",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Despesa criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "expense": {
                                            "$ref": "#/components/schemas/Expense"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/expenses/{id}": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Detalha uma despesa do usuário",
                "description": "GET /expenses/{id}\nDetalha uma despesa específica do usuário.",
                "operationId": "6f849abe1826fedd167ee90c75898e12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Despesa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "expense": {
                                            "$ref": "#/components/schemas/Expense"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrada ou de outro usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Edita uma despesa do usuário (qualquer campo)",
                "description": "PUT/PATCH /expenses/{id}\nEdita qualquer campo de uma despesa existente do usuário.",
                "operationId": "689a52d6492a9af2db6e8fb8d5e7311b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Despesa atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "expense": {
                                            "$ref": "#/components/schemas/Expense"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrada ou de outro usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Exclui uma despesa do usuário",
                "description": "DELETE /expenses/{id}\nExclui uma despesa do usuário. Responde 204 No Content.",
                "operationId": "4770780c5882debf9803f0171168fbf1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Excluída (sem corpo)"
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrada ou de outro usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Edita uma despesa do usuário (alias PATCH de PUT)",
                "operationId": "170b1c7562b046dc5a4c4d3a6d9c3650",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Despesa atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "expense": {
                                            "$ref": "#/components/schemas/Expense"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrada ou de outro usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "title": "Erro genérico",
                "properties": {
                    "name": {
                        "description": "Schema OpenAPI de erro genérico (401, 404, etc.).",
                        "type": "string",
                        "example": "Not Found"
                    },
                    "message": {
                        "type": "string",
                        "example": "Despesa não encontrada."
                    },
                    "status": {
                        "type": "integer",
                        "example": 404
                    }
                },
                "type": "object"
            },
            "ExpenseInput": {
                "title": "Entrada de despesa",
                "required": [
                    "description",
                    "category",
                    "amount",
                    "expense_date"
                ],
                "properties": {
                    "description": {
                        "description": "Schema OpenAPI do corpo de criação/edição de despesa.",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Almoço no centro"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "alimentacao",
                            "transporte",
                            "lazer"
                        ],
                        "example": "alimentacao"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0.01,
                        "example": 42.9
                    },
                    "expense_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-07-10"
                    }
                },
                "type": "object"
            },
            "Expense": {
                "title": "Despesa",
                "properties": {
                    "id": {
                        "description": "Schema OpenAPI da despesa (resposta). Classe apenas \"porta-atributos\".\nCada schema fica em seu próprio arquivo por causa do autoload PSR-4 usado\npelo swagger-php ao analisar as classes.",
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "type": "string",
                        "example": "Almoço no centro"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "alimentacao",
                            "transporte",
                            "lazer"
                        ],
                        "example": "alimentacao"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 42.9
                    },
                    "expense_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-07-10"
                    },
                    "created_at": {
                        "type": "integer",
                        "example": 1752969600
                    },
                    "updated_at": {
                        "type": "integer",
                        "example": 1752969600
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "title": "Paginação",
                "properties": {
                    "page": {
                        "description": "Schema OpenAPI do bloco de paginação da listagem.",
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 2
                    },
                    "page_count": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "Erro de validação (422)",
                "properties": {
                    "name": {
                        "description": "Schema OpenAPI da resposta de erro de validação (HTTP 422).",
                        "type": "string",
                        "example": "Unprocessable Entity"
                    },
                    "message": {
                        "type": "string",
                        "example": "Os dados enviados são inválidos."
                    },
                    "status": {
                        "type": "integer",
                        "example": 422
                    },
                    "errors": {
                        "description": "Mapa campo -> lista de mensagens",
                        "type": "object",
                        "example": {
                            "category": [
                                "Categoria inválida. Use: alimentacao, transporte, lazer."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}